clean up varbinview array module
Also changes the export pattern for `BinaryView` and related types.
