+++ vortex-array/src/arrays/varbinview/compute/take.rs
@@ -8,7 +8,8 @@ use vortex_buffer::Buffer;
 use vortex_dtype::match_each_integer_ptype;
 use vortex_error::VortexResult;
 
-use crate::arrays::{BinaryView, VarBinViewArray, VarBinViewVTable};
+use crate::arrays::binary_view::BinaryView;
+use crate::arrays::{VarBinViewArray, VarBinViewVTable};
 use crate::compute::{TakeKernel, TakeKernelAdapter};
 use crate::vtable::ValidityHelper;
 use crate::{Array, ArrayRef, IntoArray, ToCanonical, register_kernel};
