+++ vortex-array/src/arrays/varbinview/vtable/operations.rs
