+++ vortex-array/src/pipeline/types.rs
@@ -7,7 +7,7 @@ use vortex_dtype::half::f16;
 use vortex_dtype::{DType, NativePType, PType};
 use vortex_error::vortex_panic;
 
-use crate::arrays::BinaryView;
+use crate::arrays::binary_view::BinaryView;
 
 /// Defines the "vector type", a physical type describing the data that's held in the vector.
 ///
